Starting Linux Support
Started workign with command line cmake, which is super easy to get things done, maybe generate an executable that just does cmake and put the defines in the cmakelist instead of running commandline argument directly

Most of the compiler issues are solved, now having a couple linker errors I do not understand:

/usr/bin/ld: CMakeFiles/Temp_Engine.dir/EngineCode/helpers/LogHandler.cpp.o: in function `LOGGER::DumpData()':
/home/marcfly/Documents/TFG-TempEngine/Temp_Engine/EngineCode/helpers/LogHandler.cpp:89: undefined reference to `std::experimental::filesystem::v1::create_directory(std::experimental::filesystem::v1::__cxx11::path const&)'
/usr/bin/ld: CMakeFiles/Temp_Engine.dir/EngineCode/helpers/LogHandler.cpp.o: in function `std::experimental::filesystem::v1::__cxx11::path::path<char [5], std::experimental::filesystem::v1::__cxx11::path>(char const (&) [5])':
/usr/include/c++/9/experimental/bits/fs_path.h:212: undefined reference to `std::experimental::filesystem::v1::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: CMakeFiles/Temp_Engine.dir/EngineCode/helpers/SimpleTasker.cpp.o: in function `std::thread::thread<void (&)(), , void>(void (&)())':
/usr/include/c++/9/thread:126: undefined reference to `pthread_create'
collect2: error: ld returned 1 exit status
make[2]: *** [CMakeFiles/Temp_Engine.dir/build.make:174: Temp_Engine] Error 1
make[1]: *** [CMakeFiles/Makefile2:76: CMakeFiles/Temp_Engine.dir/all] Error 2

+++ Temp_Engine/EngineCode/modules/ModuleTime.cpp
@@ -46,7 +46,7 @@ float ModuleTime::CalcDT()
     float last_frame = 1000*timer.Read_mS();
     timer.Start();
 
-    engine_dt = (last_frame < max_dt) ? last_frame : max_dt;
+    engine_dt = (last_frame < max_dt) ? last_frame : (float)max_dt;
     game_dt = engine_dt * control_timings[control];
 
     return engine_dt;
